requirements.txt
pytest==7.2.0
pandas==1.5.2
google-cloud-aiplatform==1.19.1
kfp==1.8.17